pyproject.toml
[tool.poetry]
name = "ds-ai-starter"
version = "0.1.0"
description = "Entorno base para proyectos de Ciencia de Datos e IA"
authors = ["Gerardo Sepúlveda gerardosepulveda25@gmail.com"]
readme = "README.md"

packages = [
    { include = "my_project", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "^1.26"
pandas = "^2.2"
matplotlib = "^3.8"
scikit-learn = "^1.5"
plotly = "^5.24"
jupyter = "^1.0"
ipykernel = "^6.29"
python-dotenv = "^1.2.1"
pydantic-ai = "^1.14.0"
pydantic = "^2.12.4"
openai = "^2.7.2"
groq = "^0.33.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8"
ruff = "^0.6"
pre-commit = "^3.8"
pytest = "^8.3"
